--- packages/dmn-editor/dev-webapp/static/index.html
@@ -1,40 +1,0 @@
-<!--
-  ~ Licensed to the Apache Software Foundation (ASF) under one
-  ~ or more contributor license agreements.  See the NOTICE file
-  ~ distributed with this work for additional information
-  ~ regarding copyright ownership.  The ASF licenses this file
-  ~ to you under the Apache License, Version 2.0 (the
-  ~ "License"); you may not use this file except in compliance
-  ~ with the License.  You may obtain a copy of the License at
-  ~
-  ~  http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing,
-  ~ software distributed under the License is distributed on an
-  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  ~ KIND, either express or implied.  See the License for the
-  ~ specific language governing permissions and limitations
-  ~ under the License.
-  -->
-
-<!DOCTYPE html>
-<html lang="en-US">
-  <head>
-    <meta charset="utf-8" />
-    <title>DMN Editor - Dev webapp</title>
-    <meta id="appName" name="application-name" content="DMN Editor" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="icon" type="image/svg+xml" href="./favicon.ico" />
-    <!-- Issue workaround for React v16. -->
-    <script>
-      // See https://github.com/facebook/react/issues/20829#issuecomment-802088260
-      if (!crossOriginIsolated) SharedArrayBuffer = ArrayBuffer;
-    </script>
-  </head>
-
-  <body>
-    <noscript>Enabling JavaScript is required to run this app.</noscript>
-    <div id="dev-webapp" style="height: 100vh"></div>
-    <script src="index.js"></script>
-  </body>
-</html>
