+++ packages/dmn-editor/tests/e2e/connections/createNewNodeFromExistingOne.spec.ts
@@ -23,12 +23,12 @@ import { env } from "../../../env";
 test.beforeEach(async ({ page, browserName }, testInfo) => {
   test.skip(browserName === "webkit", "hover() and dragTo() cause troubles on webkit");
   await page.goto(
-    `http://localhost:${env.dmnEditor.storybook.port}/iframe.html?args=&id=example-dmndevwebapp--empty-model&viewMode=story`
+    `http://localhost:${env.dmnEditor.storybook.port}/iframe.html?args=&id=dev-web-app--empty-model&viewMode=story`
   );
 });
 
 test.describe("Create new node from existing one", () => {
-  test("InputData -> Decision", async ({ page }) => {
+  test.only("InputData -> Decision", async ({ page }) => {
     await page
       .getByTitle("Input Data", { exact: true })
       .dragTo(page.getByTestId("rf__wrapper"), { targetPosition: { x: 100, y: 300 } });
